Add feature flag to cleanup

+++ src/main/resources/tasks/deleteExpiredEventLog/deleteExpiredEventLog.ts
@@ -1,7 +1,11 @@
 import { deleteExpiredEventLogsForQueries, deleteExpiredEventLogsForJobs } from '/lib/ssb/cron/eventLog'
+import { isEnabled } from '/lib/featureToggle'
 
 export function run(): void {
   log.info(`Run Task: deleteExpiredEventLog ${new Date()}`)
   deleteExpiredEventLogsForQueries()
-  deleteExpiredEventLogsForJobs()
+
+  if (isEnabled('delete-joblog-jobs', false, 'ssb')) {
+    deleteExpiredEventLogsForJobs()
+  }
 }
